# [core] Lazy import modules to reduce importing time