--- a/sky/adaptors/common.py
+++ b/sky/adaptors/common.py
@@ -6,18 +6,19 @@ from typing import Any, Callable, Optional, Tuple
 
 
 class LazyImport:
-    """Lazy importer for heavy 3rd party modules or cloud modules
-       only when enabled.
+    """Lazy importer for modules.
 
-    We use this for heavy 3rd party modules as they can be time-consuming to
-    import, for example, numpy(700+ms), pendulum(500+ms), cryptography(500+ms),
-    pandas(200+ms), and networkx(100+ms), etc. With this class, we can avoid the
-    unnecessary import time when the module is not used (e.g., `networkx` should
-    not be imported for `sky status` and `pandas` should not be imported for
-    `sky exec`).
+    This is mainly used in two cases:
+    1. Heavy 3rd party modules: They can be time-consuming to import
+    and not necessary for all `sky` imports, e.g., numpy(700+ms),
+    pendulum(500+ms), cryptography(500+ms), pandas(200+ms), and
+    networkx(100+ms), etc. With this class, we can avoid the
+    unnecessary import time when the module is not used (e.g.,
+    `networkx` should not be imported for `sky status` and `pandas`
+    should not be imported for `sky exec`).
 
-    We also use this for cloud adaptors, because we do not want to import the
-    cloud dependencies when it is not enabled.
+    2. Cloud modules in cloud adaptors: cloud dependencies are only required
+    when a cloud is enabled, so we only import them when actually needed.
     """
 
     def __init__(self,