--- a/sky/backends/backend_utils.py
+++ b/sky/backends/backend_utils.py
@@ -17,6 +17,7 @@ from typing import Any, Dict, List, Optional, Sequence, Set, Tuple, Union
 import uuid
 
 import colorama
+import filelock
 from packaging import version
 from typing_extensions import Literal
 
@@ -51,7 +52,6 @@ from sky.utils import timeline
 from sky.utils import ux_utils
 
 if typing.TYPE_CHECKING:
-    import filelock
     import requests
     from requests import adapters
     from requests.packages.urllib3.util import retry as retry_lib
@@ -64,7 +64,6 @@ if typing.TYPE_CHECKING:
     from sky.backends import local_docker_backend
 else:
     yaml = adaptors_common.LazyImport('yaml')
-    filelock = adaptors_common.LazyImport('filelock')
     requests = adaptors_common.LazyImport('requests')
     rich_progress = adaptors_common.LazyImport('rich.progress')
     adapters = adaptors_common.LazyImport('requests.adapters')