--- a/sky/jobs/scheduler.py
+++ b/sky/jobs/scheduler.py
@@ -41,16 +41,22 @@ import contextlib
 from functools import lru_cache
 import os
 import time
-
-import filelock
-import psutil
+import typing
 
 from sky import sky_logging
+from sky.adaptors import common as adaptors_common
 from sky.jobs import constants as managed_job_constants
 from sky.jobs import state
 from sky.skylet import constants
 from sky.utils import common_utils
 from sky.utils import subprocess_utils
+
+if typing.TYPE_CHECKING:
+    import filelock
+    import psutil
+else:
+    filelock = adaptors_common.LazyImport('filelock')
+    psutil = adaptors_common.LazyImport('psutil')
 
 logger = sky_logging.init_logger('sky.jobs.controller')