--- a/sky/jobs/utils.py
+++ b/sky/jobs/utils.py
@@ -16,14 +16,13 @@ import typing
 from typing import Any, Dict, List, Optional, Set, Tuple, Union
 
 import colorama
-import filelock
-import psutil
 from typing_extensions import Literal
 
 from sky import backends
 from sky import exceptions
 from sky import global_user_state
 from sky import sky_logging
+from sky.adaptors import common as adaptors_common
 from sky.backends import backend_utils
 from sky.jobs import constants as managed_job_constants
 from sky.jobs import scheduler
@@ -40,8 +39,14 @@ from sky.utils import subprocess_utils
 from sky.utils import ux_utils
 
 if typing.TYPE_CHECKING:
+    import filelock
+    import psutil
+
     import sky
     from sky import dag as dag_lib
+else:
+    filelock = adaptors_common.LazyImport('filelock')
+    psutil = adaptors_common.LazyImport('psutil')
 
 logger = sky_logging.init_logger(__name__)