--- a/sky/provision/kubernetes/network_utils.py
+++ b/sky/provision/kubernetes/network_utils.py
@@ -1,19 +1,25 @@
 """Kubernetes network provisioning utils."""
 import os
 import time
+import typing
 from typing import Dict, List, Optional, Tuple, Union
-
-import jinja2
-import yaml
 
 import sky
 from sky import exceptions
 from sky import sky_logging
 from sky import skypilot_config
+from sky.adaptors import common as adaptors_common
 from sky.adaptors import kubernetes
 from sky.provision.kubernetes import utils as kubernetes_utils
 from sky.utils import kubernetes_enums
 from sky.utils import ux_utils
+
+if typing.TYPE_CHECKING:
+    import jinja2
+    import yaml
+else:
+    jinja2 = adaptors_common.LazyImport('jinja2')
+    yaml = adaptors_common.LazyImport('yaml')
 
 logger = sky_logging.init_logger(__name__)