<!-- Describe the changes in this PR -->
Part of #4678 

Improvements for the core operations:
- sky exec, mean time for 5 runs from 3.45s to 2.747s, about 20% faster
- sky launch on existing cluster, mean time for 5 runs from 14.901s to 13.08s, about 12% faster
- sky status, mean time for 5 runs from 2.426s to 2.157s, about 11% faster
- sky queue, mean time for 5 runs from 3.155s to 2.734s, about 13% faster

This PR only covers the third-party modules to import lazily, however, from the test result below, it still takes a long time to import the whole sky module.
So this needs to be improved further, e.g.:
- Lazy import of other unnecessary modules
- Split some core modules further to make the import more fine-grained


<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [x] Any manual or new tests for this PR (please check the separate comments below)
- [ ] All smoke tests: `pytest tests/test_smoke.py` 
- [ ] Relevant individual smoke tests: `pytest tests/test_smoke.py::test_fill_in_the_name` 
- [ ] Backward compatibility tests: `conda deactivate; bash -i tests/backward_compatibility_tests.sh`