@DanielZhangQD 2025-02-23
Method to verify the import time:
Execute the [scripts](https://gist.github.com/DanielZhangQD/0db7b830a9a9aa3e346d0d57732b661d) with command:
```
./run_clean_import.sh --force-reinstall
```

Result of this change:
- MacOS with arm: total time from 5.258s to 2.053s
  - Master
<img width="1160" alt="image" src="https://github.com/user-attachments/assets/32e8449c-d804-47bc-983a-1ea6e5ed89f9" />

  - This PR
<img width="1174" alt="image" src="https://github.com/user-attachments/assets/7cebbabb-b094-4678-afb9-20a72d4b4ff8" />

- Linux: total time from 1.595s to 0.916s
  - Master
<img width="1095" alt="image" src="https://github.com/user-attachments/assets/7383c230-657c-4a1a-a038-d63da5412c50" />

  - This PR
<img width="1097" alt="image" src="https://github.com/user-attachments/assets/9dc8c2b3-74f5-43f1-9212-fbd823b2691c" />