@DanielZhangQD 2025-02-23
Hi @Michaelvll, could you please help trigger smoke_test for this PR? Thanks!

BTW, as I have described in the PR description, this PR only covers the 3rd party modules and the import time is still long and needs to be reduced further.
However, further improvement may need to touch the business logic of the core modules, e.g. split some complicated modules, and I prefer to do that in separate PRs later, what do you think?