@Michaelvll 2025-02-23
Thanks for submitting this PR @DanielZhangQD! Doing the core part inanother PR sounds good to me. 

I have some quick questions:
1. The main goal of doing this is to speed up some of our core operations. Does making this lazy import help: (a) `sky status`, (b) `sky launch`, (c) `sky exec`, (d) `sky queue`? Since some of those operations involves `ssh` and run inline python script that imports `sky`, reducing the import time may help those case. It would be better to have some profiling for this.
2. Is there any drawback for making those third-party lazy import?