@DanielZhangQD 2025-02-24
Hi @Michaelvll, thanks for the info!
I will:
1. Profile those core operations
2. Yes, there may be some potential issues for the lazy import, e.g.:
- 2.1 Race condition in multithread importing
- 2.2 Potential broken type hints
- 2.3 Missing or incompatible dependencies are only detected at runtime when the lazy import is triggered, not during startup
- 2.4 The first use of a lazily imported module incurs import overhead
- ...

For 2.1, I see `Class LazyImport` already includes a lock to avoid this issue. 
For the others, we must thoroughly test this PR to avoid functionality breaks and balance the benefits v.s. cost

And I will fix the issue with quicktest-core.